Fix a bug where clicks were being triggered on disabled elements
Have no idea why this happens, but it is clearly caused by the previous
commit.

It seems that AuthPhoneNumber.tsx has to be imported in order for app
to work correctly. This component is not used anymore, so it must be
something outside the component in the file that is causing this.